+++ api/Controllers/StockController.cs
@@ -21,6 +21,8 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stocks = await _stockRepo.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto()).ToList();
 
@@ -31,6 +33,8 @@ namespace api.Controllers
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await _stockRepo.GetByIdAsync(id);
 
             if (stock == null)
@@ -43,6 +47,8 @@ namespace api.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequest request)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = request.ToStockFromCreateDto();
 
             await _stockRepo.CreateAsync(stockModel);
@@ -52,9 +58,11 @@ namespace api.Controllers
 
         // UPDATE
         [HttpPut]
-        [route("{id:int}")]
+        [Route("{id}")]
         public async Task<IActionResult> Update(int id, [FromBody] UpdateStockRequest request)
 {
+                if (!ModelState.IsValid)
+                    return BadRequest(ModelState);
     var stockModel = new Stock
     {
         Symbol = request.Symbol,
@@ -71,10 +79,13 @@ namespace api.Controllers
     return Ok(updatedStock.ToStockDto());
 }
         // DELETE
-        [HttpDelete
+        [HttpDelete]
         [Route("{id:int}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var stockModel = await _stockRepo.DeleteAsync(id);
 
             if (stockModel == null)
